border: check errors before updating metadata
Check for possible errors when incrementing the path before updating the
packet metadata, which can result in an invalid state.

The issue showed up when the border router processes a BadSegment
packet. Because the metadata is updated, when we create the scmp reply
and reverse the packet, we fail to increment the reversed path, not
being able to send the error back.

With this change, the packet metadata is only updated once we have
verified that there are no errors when incrementing the path.