border: best effort approach for scmp error quotes
The border router findL4() method keeps track of the last parsed
header/eextension.

With this approach when generating SCMP quotes, it includes as much
relevant data as possible.

If the border router failed to parse the extensions, thus could not find
the L4 offset, it quotes everything after the last parsed extension as
the L4 quote.

If it found the L4 header but it is an unknown/unsupported protocol,
then it quotes everything from the L4 offset as the L4 quote.

With a best effort approach, if the buffer where the Payload is written
to does not have enough space, it will fail and it is up to the caller
to modify the SCMP quotes to fit in the buffer.