Reduce raciness in test
fetchSegsFromDBRetry select on ctx.Done() and time.After().
If the setup/calling of ctx.Done() takes more than what we pass in time.After,
it can be that both channels are ready at the same time and then the test might fail.

By increasing the timeout in the test this should no longer be a problem.