Process SCMP errors caused by SCMP packets (#2052)
There are a few issues when processing SCMP errors caused by SCMP packes,
like SCMP echo, by the border router and the dispatcher:
The border router fails to include the SCMP Info field, containing the
ID which is required to deliver the packet to the application.
The dispatcher does not process SCMP errors of SCMP packets.

This change adds the required support to the border router and the
dispatcher as well as modifying the SCMP tool to be able to deal with
SCMP revocations. It also adds a refresh flag to the scmp tool, which
now talks to sciond instead of the path manager.