border: fix error replies for unsupported L4 protocols
When generating SCMP errors replies for bad packets, the border router
quotes different parts of the bad packet, this quotes also depend on how
much the border router is able to parse of it.

The issue shows with unsupported procotols, ie. TCP, where the border
router generates an SCMP error reply with Meta L4Type value 0, meaning
None/HopByHop, which is obviously wrong.

The problem arise because we try to retrieve the L4 Type from the parsed
L4 header, which we do not parsed as it is unsupported.

The solution is simple as rpkt already has the L4Type.