Make snet.Addr implement JSONUnmarshaler.
In the RAINS configuration file we wanted to be able to pass in a SCION
address as a string instead of having the full expansion of this struct.
Since this configuration is automagically parsed by `encoding/json`, it
was not possible to pass a string in to the configuration file. Hence I
propose adding a JSONUnmarshaler method which calls AddrFromString.