BR: Acceptance Test for IP changing
Add acceptance tests for BR topology reloading with ip changes.
The test checks that changing the local and remote ip of an
interface works as expected.

To change the remote ip, a two step process is necessary.
This will be removed when #2254 is fixed.

Also, parts of the `topo_br_reload_if_port` acceptance test are refactored.