CS: Change lookup strategy
Instead of querying the dst CS we always query:
* the ISD-local core CS for ISD-local queries
* the remote core CS for remote-ISD queries

This should result in better caching:
* Connectivity should be cached (from beaconing)
* Trust material should be cached (was required to verify segments)

Also move CreateSign and VerifyChain functions from trust to a cs internal package.
Those functions are only used there and it is not yet sure if they make sense for other code.
Furthermore CreateSign now works against the trustDB since it should always have the information locally.
This was required to prevent calling ChooseServer on startup (otherwise we would have a dependency loop).

Fixes #2226