CS: Add a tool to load cust PK from file to trustDB
Doing this in the CS itself at startup is not always what we want.
For example during a trust reset we might not want to stop the CS but just insert the new PK.
Furthermore this only has to be done once and not on every start of the CS.

The tool expects a customer directory which contains the keys to load and a config for the DB.
For the DB connection config the CS config file can be passed.

Fixes #2307