Ringbuf: Avoid getting stuck writing/reading 0 entries
Previously, when returning a slice of length zero to the
ringbuffer using `Write` and the number of writable packets
is zero, the call is blocked until the next read happens.

Previously, when reading a slice of length zero from the
ringbuffer using `Read` and the number of readable packets
is zero, the call is blocked until the next write happens.

Now the calls follow the expected behaviour and are no
longer blocked under these conditions,.