Full-blown parsing of sequence path policies
This adds all the missing sequence grammar except & and !.
As for the two missing operators, these can't be implemented as
a simple state machine and may even make the entire thing
Turing-complete (?). Given that the most obvious use cases
for & and ! can be covered by ACLs maybe the operators
won't be needed at all.

Fixes #2338