BR: Remove HSR
The HSR code has not been touched in a long time.
It is outdated and will not work in the current state.

This PR removes all occurences of the HSR.
Registering the networking stack in setup.go is still
supported, to allow HSR be readded easily when we
want to do so.

Additionally, remove the setupNetHooks:
- remove setupNetStartHooks
- remove setupNetFinishHooks