BK: Add retry for jobs when an agent failed (#2350)
Buildkite: A job will show an exit status of -1 if the connection to the agent is lost.
When agents haven't been used for some time there can be a race condition between AWS stopping machines and a new build acquiring an agent. Then the agent may be stopped while already running a job again.