itopo: Split topo in static and dynamic
Split itopo topology into a static and dynamic version.

Currently, only the static topology can be set and fetched.
The code also contains an outline how the dynamic topology
will be used in a future PR.

The services currently only support setting the static
topology through a sighup reload. Querying the discovery
service will be added in a future PR.