braccept: revocation test and rework segment generation (#2302)
Add support and test for revocation. For such test, a Pre/Post step was
added to each test, providing the ability to send a packet before and
after (respectivily) the test itself. In the Pre step an IFStateInfo
deactiving an interface is sent to the BR, then a IFStateInfo activing
the same interface is sent in the Post, to restore the original BR
state.

Path segments are now defined as strings using an intuitive syntax,
maintaining all the previous flexibility to define path segments.

L4 UDP and SCMP constructors have also been modified to take a scion
header and payload to properly calculate the expected checksum during
building and matching of packets.

Fixes #2268