BR: Split topology in static and dynamic
The split is achieved by using the itopo package.

This change adds a border router specific validator to itopo
which enforces the mutability rules according to `itopo/doc.go`.

On sighup, the border router sets the new static topology of itopo.

Fetching from the discovery service is implmeneted later on.