Config: Establish consistency between configs
The config format has organically grown and is currently very inconsistent
throughout the code base.

This PR synchronizes the different config approaches
in one simple approach. The goals are:
1. consistency
2. large test coverage
3. Decouple state from config (i.e. Remove `env.General.Topology`)
4. fewer code duplication
5. validation framework
6. sample config generation

A good place to start for reviewing is `lib/config` to get an overall
feel for the purpose of the PR.

Config structs have to implement the `config.Config` interface.
There are three parts to this interface:
1. Initialization
2. Validation
3. Sample generation

Initialization: Each config struct is responsible for initializing its
fields. This allows hierarchical initialization. I.e. whenever the
a config struct is initialized, all child config structs are also
initialized.

Validation: Each config struct is responsible for validation its fields.
This allows for hierarchical validation, similar to initialization above.

Sample generation: Sample are hierarchically put together. A config
struct provides the `ConfigName`. This forces consistent keys across
multiple services for the same config struct.
The provided `ctx` map currently only holds the instance id. However,
through this map, additional information can be passed to the child
config structs if they will ever be necessary.

The logical flow from an applications point of view is the following:
1. Unmarshal config file into top-level config struct
2. Initialize default values for config struct
3. Setup logging early to have formatted output on errors
4. Validate config
5. Start with rest of execution

For complex applications this is expressed in the (setupBasic/setup)
pattern.
For simpler applications the (setupBasic/cfg.Validate) pattern suffices.

The hierarchical approach is also used in the unit tests.
Packages with config structs in the library must have a sub-package
with the package name and suffix `test`.
This sub-package has to files:
- config_test.go: Unit test that sample is correct for all config
structs in parent package.
- config.go: Provide the functions
`InitTest"ConfigName"`/`CheckTest"ConfigName"` that initialize the
config struct for the sample correct unit test and checks it.

Notice that this change does not address the `snake_case` vs `camelBack`
vs `CamelCase` inconsistencies. This would grow the changes even larger.
The change is inter-operable with the current config file setup.