Remove ctx from deduper.Request method
The context was shadowed in the current implementation anyway.
Furthermore it could lead to weird interactions:
if the first caller of a deduped request has a long timeout and the second one a short one
Only the timeout of the first requestor would be respected.
That is why callers of Request anyway also select on the ctx.Done() channel,
thus we don't have to pass it to the Request method.