braccept: rework test generation
Main changes:
- Syntax to describe packets, including custom parsing logic
- Generic layer comparison.
- Each test is now a Go funtion, removing test processing constraints.

In its basic form, each layers is defined by a text line following
the following sytnax:
   TAG: prop1=val1 prop2=val2

where TAG has two characteristics:
1. It specifies the type of layer.
2. It uniquely names a layer.

As an example:
   IP4: Src=192.168.12.3 Dst=192.168.12.2 NextHdr=UDP
   UDP_1: Src=40000 Dst=50000 Checksum=0

The above example defines a packet with two layers, IP4 and UDP,
with tags IP4 and UDP_1 respectively.
Tags are useful to refer to specific layers in a packet, such that
we can update only some layers from a cloned packet.

The parse is indentation aware, meaning that all layers need to have
the same tab indentation as the first layer.

It is also possible to define a layer which consists of multiple
lines:
   IP4: Src=192.168.12.3 Dst=192.168.12.2 NextHdr=UDP
   UDP: Src=40000 Dst=50000 Checksum=0
   SCION: NextHdr=UDP CurrInfoF=4 CurrHopF=5 SrcType=IPv4 DstType=SVC
       ADDR: SrcIA=1-ff00:0:2 Src=172.16.2.1 DstIA=1-ff00:0:1
       Dst=192.168.0.51
       IF_1: ISD=1 Hops=3 Flags=Peer
           HF_1: ConsIngress=261 ConsEgress=0   Flags=Xover
           HF_2: ConsIngress=211 ConsEgress=151 Flags=Xover
           HF_3: ConsIngress=0   ConsEgress=621
   UDP: Src=1111 Dst=2222

From a parsing point of view, every line with deeper indentation
than the layer itself belong to the layer, so in the example above,
the SCION layer would consist of:
   SCION: NextHdr=UDP CurrInfoF=4 CurrHopF=5 SrcType=IPv4 DstType=SVC
       ADDR: SrcIA=1-ff00:0:2 Src=172.16.2.1 DstIA=1-ff00:0:1
       Dst=192.168.0.51
       IF_1: ISD=1 Hops=3 Flags=Peer
           HF_1: ConsIngress=261 ConsEgress=0   Flags=Xover
           HF_2: ConsIngress=211 ConsEgress=151 Flags=Xover
           HF_3: ConsIngress=0   ConsEgress=621

This rework also changes the way layers are compared against expected
packets in a much more generic way by doing bytes and string comparision.
It also uses a string diff package that colors differences,
make it much easier to spot the issues between the received and expected
packet.

Fixes #2579