Healthpool: Add general healthpool implementation  (#2487)
The package `healthpool` provides a generic way to keep track of the health infos for a set of keys.
It is used to implement other health pools for specific purposes. They will be implemented in sub-packages. Client packages will use these more specific implementations.

E.g. `svcinstace.Pool` will implement such a specific health pool that keeps track of the health of all service instances for a given service type in the local AS. When it is implemented, `lib/discovery/InstancePool` can be replaced.

This also ties in with #2434