Mutiple changes to test approach
* We fixed the mac addresses of the interfaces we create, so we do not
need to create temp file with such information, etc.

* Change in topology and test approach in general. The main idea is
to reuse tests between different configurations.
Thus, we now have a custom topology for each configuration where the
BR under test is always brA/core-brA in AS ff00:0:1.

The result is that the same tests that run on a BR with a single
external interface (child/parent/peer/core) can also be run on
a BR with multiple interfaces.