BS: Add propagator
This change adds a beacon propagator.
The propagator fetches the beacons to propagate from the beacon
provider and sends them on all active applicable interfaces.

In a core BS, these are the core links.
In a non-core BS, these are child links.

In case the beacon would create a loop on a given interface, the beacon
is not forwarded to that interface.

fixes #2563