BS: Add topology provider
This PR adds the topology provider to decouple the tasks from the
itopo package. This makes tests cleaner and and overall improves the
code quality. If this shows to be effective, we can expand this pattern
to other parts of the code base.

Additionally, the initialization of tasks is refactored to allow
simpler extensibility and reduce constructor signature.