CS: Force cancel core pusher
Before, when shutting down the certificate server with the core pusher
running and the certificate server not having paths to the core would
result in the certificate server to wait for up to a minute.

With this PR the CS kills the task instead of waiting for it to finish.

Additionally, the SegPusher now uses go routines to avoid head of line
blocking.