Fix mangling unit tests
Some unit tests mangle bytes to cause invalid input.
A few do so by assigning a constant to a given byte, resulting in a
non-zero chance that the byte equals the initial one and is not altered.
This PR ensures that the bytes are always mangled to avoid unnecessary
probabilistic unit test failures.