Ordered topology start
This PR first starts the dispatcher and border routers to ensure
initial connectivity.

Currently, some tests fail because the beacon server are started
before the border routers and the ifid keepalives are not delivered.
By the time the border router has started, the interface is already
revoked. The introduced ordering will avoid this problem.