PS: Request missing TRC from core CS (#2723)
The path server determines whether an AS in a remote ISD is core or
not using the TRC. If the TRC is missing locally, the path server
currently queries the local certificate service for the TRC.

In a non-core AS, the certificate server discovers TRCs for remote ISDs
through the path server. Thus, it is highly unlikely that it already has
the TRC. This causes the certificate server to request a path to the
local core from the path server, leading to many rpc calls and a loop
in addressed services.

With this change, a path server in a non-core AS immediately fetches
the TRC from a local core certificate server to avoid rpc calls
and reduce latency.