Remove all ineffective assignment
Sprinkled through the codebase are ineffective assignments.
Most of them are in tests that could result in passing tests with error.

The change in `go/lib/pktcls/pred_ipv4.go` is an actual bug, as the
error is not handled. It is not caught by bazel, since the code is
not built.

Additionally, refactor the portlist_test to use require.