Add SegReplyHandler for seg verification and storage
This abstract partial logic of fetcher from sciond to its own lib.
The idea is that this lib can also be used in the PS.
The SegReplyHandler verifies and stores a segment reply.
It can be triggered to store a partial result earlier and
it will provide feedback how many segments have been stored early.
This way users of the lib can potentially reply to client before they have processed the full reply.