Add make target to generate go/proto files
Generate *.capnp.go and structs.gen.go files in bazel and then copy to
source directory in a make target.

For the case of structs.gen.go which is generated with a simple shell
script, this setup is somewhat overcomplicated.
Generating the *.capnp.go files does use bazel infrastructure to get the
capnp tooling.
To keep things symmetric, both cases are handled the same.