infra: Refactor TrustStore interface
This PR refactors the infra.TrustStore interface.

Different client packages care about different personalities of the
trust store. Many TRC retrievals in the code base are simply there to
check whether an AS is core or not.

The goal of this refactor is to abstract interactions with the CP-PKI
such that regular infra services do no longer have to care about the
crypto material. This will make the transition to the new formats for
certificate chains and TRCs a lot simpler.

The certificate server will need an extended TrustStore interface, since
the interacts are not limited to sign/verify message and resolve core.