Add method to filter segments with a path policy
This method pre-filters a list of segments with a given path policy.
Note that filtering with sequences is not easily possible since we do not know where a segment would fit into the sequence (up, core, down).
A possibility to add sequence filtering for segments would be to split the sequence into up, core, and down parts.

Note that filtering with ACLs should only be done on core segments.
Up and down segments contain peering interfaces which might be used in the final path, eventhough the ACL would filter it.

Fixes #2820