topology: fix panic and make overBindAddr public
Both PublicOverlay and BindOverlay methods on TopoBRAddr would panic
there is no address set for the requested overlay type.

This change fixes that and also makes OverBindAddr a public type so it
is easier to work with it as the only way to get such objects is by
parsing a topology file.