SPKI: Generate prototype TRC
This change allows generating prototype TRC payloads.
A prototype TRC will be signed and then combined to a signed TRC.
Signing and combining is left for a future PR.

For convenience, a 'gen' command will also be added in the future that
executes the three phases in a single call.

contributes to #2982