serrors: Add serrors package
This should in the long term replace:
-errors.New
-common.NewBasicError

The package is ready to be used with Go1.13 (or xerrors) Is and As.

Also it provides all the functionality that was provided by common.BasicError/common.MultiError.

By putting it in a serrors package we can make the API slightly nicer to use.