TrustStore: Define components
This PR introduces the different components of the trust store.
In the end, the client package facing parts of the trust store are
constructed out of the `CryptoProvider` and the `Inspector`.

The other components are used internally to modify the behavior of the
trust store.

E.g. different `Recurser`s will be used between the CS and sciond.

The purpose of the components is the following:
- Inspector: Implement the infra.ASInspector
- CryptoProvider: Main logic when fetching trust material. It decides
  when to query the database and/or the network.
- Inserter: Verify and insert TRC/certificates into the database.
- Recurser: Decide whether recursion is allowed.
- Router: Get path based on the trust material subject.
- RPC: Abstraction of the messenger stack.