TrustStore: Implement crypto provider (#3149)
The crypto provider handles crypto material requests. If the
requested crypto material is in the database, it is returned.
Otherwise, the crypto provider decides based on the recurser
whether it can issue a network request. If it is allowed to,
it selects the path and remote using the router and calls
resolver to resolve all the missing material.

This PR is the start of a series of PRs that make the new trust store
operational. The individual components are then glued together into
the trust store. To the client package, the trust store appears as
a single unit.

Contributes to #3115