segfetcher: Classify error better
This partially classifies the error of the segfetcher.
To do it better we would have to use classified errors on all used libs,
which would be a huge amount of work.
And it's not yet sure that this approach is the best for metrics.
Maybe individual error counters make more sense.

Fixes #3240