SPKI: Prototype TRC generation
Add support for generating prototype TRCs based on the TRC configuration
files.

The version to generate can be provided via command line flag. If no
version is specified, the tool searches for the newest TRC configuration
file and uses its version.

This change also removes the old code for generating TRCs for a clean
slate approach.