Avoid vendored deps to simplify go.mod compatibility
Vendored dependencies are automatically used during builds with bazel.
These dependencies are not visible when generating a go.mod file from
WORKSPACE.
As go with modules will ignore the vendored dependencies, the result
will be that the latest versions of these inderect dependencies will be
fetched. This is bad and will typically not work.