Allow unspecified local address in snet Dial/Listen
Register unspecified address (i.e. 0.0.0.0) in the dispatcher.
The godispatcher already handles this correctly.

When using an unspecified source address for sending a packet, look up
the route to the next hop in the kernel routing table and use the
corresponding source address as source address in the SCION packet.

Note on AppAddr.Copy:
We want to be able to copy a non-nil but uninitialized snet.Addr.
Previously this failed when copying nil *AppAddr member "Host". Add nil
check in AppAddr.Copy, as is done in other places.