Add snet.{SVC,UDP}Addr to replace snet.Addr
In this PR the `snet.Addr` is being deprecated and being replaced with
two new ones `snet.{SVC,UDP}Addr`. This is required because we want to
remove the usage of `addr.AppAddr` overall in our code base and make the
API towards SVC or UDP scion addresses cleaner.

As from now, `snet.Addr` should not be used. In the code the `net.Addr`
interface should be the way to communicate between packages, and a
switch statement should always take place.

For example
```go
//  public API should ideally use net.Addr not the type
func MyPublicRead( ..., address net.Addr, ...) (int,err){
	switch a:=a.(type) {
        case *snet.Addr:
	     panic("we should never use that")
 	case *snet.UDPAddr:
	     doStuffWithUDP(a)
	case *snet.SVCAddr:
         doStuffWithSVC(a)
    default:
	     return nil, fmt.Errorf("unsupported type  a(%T)", a, ,a)
}

// not public API should ideally use the type not the addr
func doStuff(a snet.UDPaddr) {
  if a.IA == mylocal IA { ...}
}
```

Currently `{SVC,UDP}addr` are using `snet.Addr` because they share
functionality (same fields, some common functions but in the future that
will change.

In this PR:
    - modified BS,PS,CS to use new addrs type
    - messenger to support  new addrs
    - make internals to work with SVCAddr if possible
    - Copy the path when creating the legacy address
    - Always print the acceptance log line, needed for tests

Contributes scionproto/scion#3344