Fix IPv6 registrations in godispatcher (#3451)
The godispatcher uses two separate port tables for IPv4 and IPv6
addresses. While the lookup has selected the applicable table based on
the address type, the registration was always using the table for IPv4.

Also enhance the tests for UDPTable to check that a lookup actually
succeeds after a registration.