CPPKI: Do not double base64 encode
We use the JWS standard that prescribes that the payload and protected
metadata must be encoded as base 64 raw url safe encoding.

In the go code, we have typed byte slices with respective encode and
decode functions because the json library only allows using the standard
base64 encoding.

Helpfully, when calling json.{Unm,M}arshal, byte slices are encoded in
base64 automatically. Which results in double encoding in our case.

This PR changes the encoded types to string types to prevent the
json library from encoding the data.