Delete old crypto code move v2 inplace
This moves the new TRC, Cert v2 code into the place of the old code.
Also moves v2 in the scion-pki to the parent folder.

This PR deletes cryptosyncer of the PS and custpk-load tool, because they used the old API.
Those removed processes have to be rewritten against the new version of the code.