PS: Remove request holding
Holding requests in the PS can lead to a lot of outstanding requests
if a certain destination AS is not running.

The holding of requests helped to make integration tests more stable,
because at the start of a full topology it is true that segments will be registered in the future,
in general this does not hold.

For production code this can lead to a lot of CPU usage for nothing.