Fix connected Write on snet.Conn
Minimal fix for `snet.Conn.Write`.
Previously, passing `nil` into `WriteTo` would automatically pick the
connected remote address, if available. With some of the previous
refactoring (amongst other, in #3571), this is no longer the case. Right
now, `Write` is simply broken.
Apply the simplest fix, just specify to use the connected address
(`base.remote`) in `Write`. Note that the behaviour is still subtly
different than it used to be (e.g. no error if WriteTo is invoked on a
connected socket), but this seems to be the accepted behaviour change
within the refactoring in #3571.

Additionally, add a specific error message for the `nil` address case
in `WriteTo`, as "Unable to write to non-SCION address" was misleading.