Monolithic Control Service
Also:
- Micro service support was removed from the topology generator;
- In monolith, beacons are now written to path db directly instead of
through intra-AS RPCs;
- In monolith, routing is done directly on top of the path database,
instead of going through SCIOND;
- Move SCIOND Host API to TCP; this allows SCIOND to run on a
different machine;
- The sd.Reliable, sd.Unix, sd.SocketFileMode, sd.DeleteSocket, and
sd.Public SCIOND config options have been removed. Current or later
SCIOND implementations will ignore these fields if they are set.
- The sd.address config option has been changed. It now also specifies
the TCP address of the SCIOND Host API.
- SCIOND no longer creates socket folders (e.g., in /run/shm) when
running.
- Automatically finding a local-machine SCIOND depending on AS number is
no longer possible. This was used by a few tests and that run multiple
ASes on the same machine. It is now required to specify which SCIOND to
use.
- The topology generator generates a file containing the list of SCIONDs
for every AS in `./sciond_addresses.toml`. This is used by some tests to
find the SCIOND servers.
- pingpong, scmp, showpaths: scionFromIA command line flag removed;
sciond command line flag reworked to point to the TCP address of SCIOND.
- SCIOND now talks to the monolith over TCP instead of QUIC;

Co-authored-by: Lukas Vogel <lukedirtwalker@gmail.com>
Co-authored-by: Sergiu Costea <sergiu.costea@gmail.com>