Guard against system clock moving backwards.
The timestamp returned from the kernel is in system time. This means
it's possible for the kernel received time to appear after the
application receives the packet, which then can cause logic errors to
occur in code that depends on the ordering of these timestamps.

Also guard against this in the BR's packet processing time metric
calculation, as system time could move backwards during the processing
of a packet.