integration: speed up dockerized end2end test
The overhead of running a binary in a docker container with
docker-compose exec is around 1 second.

Since our integration tests are serialized and a cartesian product between
all ASes is taken, this overhead easily amounts to multiple minutes.

With this PR, all servers are started in parallel for the end2end test.

All pairs with the same source are executed in on single docker exec
command. And all executions are done in parallel.

To get a progress update, the end2end binary talks through a unix socket
to the integration binary.

We can eventually move the server wait to this RPC framework as well.

With this style of invoking the end2end test, we can reduce the execution
time on CI from >10 minutes (timeout) to a bit more than 1 minute.
Most of the execution time is actually spent in starting the topology,
not in execution the tests.