ci: attempt to make integration tests more reliable
It looks like CI is not capable of handling large parallelism in the dockerized topology.
Until we have time to investigate this, let's just reduce the parallelism to 1 on CI.

Also, increase the timeouts for the integration tests, because it can take quite a bit of time for containerized applications to start with docker-compose.