scion: one command to rule them all (#3708)
Cobra has shown to be a very useful framework for writing CLI applications.
It makes it easy to write applications with many sub-commands and high configurability.

We currently have many binaries in the code base, that do not really warrant to be their own application. With the approach that we put most application logic in `go/pkg/...`, we can easily create standalone binaries (when needed/desired) and also bundle them in an all-in-one scion command.

This PR is a exploration how this could look like.
At the base is the `scion` command that bundles all the other commands.

For this exploration, I moved the showpaths logic to `go/pkgs/showpaths` such that it can be imported as a library. The old `showpaths` binary is not affected by the change, and if we continue to support the standalone binary, this will be easy to do.

Under `go/scion` the all-in-on binary is set up. Currently, it only has 3 commands:
- `completion` : generate auto-completion
- `version`: display the scion version
- `showpaths`: run the showpaths application

The nice thing is, that we gradually can add sub-commands that make sense to be bundled.
E.g., I envision `scion echo`, `scion traceroute`, `scion pingpong`, and the soon to be added `scion pathdbdump` (#3707) tools to be part of `scion` eventually.